cleanup: remove unnecessary stability checks
